% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlingeo-package.R
\docType{package}
\name{berlingeo-package}
\alias{berlingeo-package}
\alias{berlingeo}
\title{Spatial Data Files for Berlin (Germany)}
\description{
This package contains spatial data files
for various and administrative boundaries of Berlin, Germany.
}
\author{
Patrick Hausmann \email{patrickhausmann@covimo.de}
}
\keyword{package}
