% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlingeo_data.R
\docType{data}
\name{lor_bzr_sf}
\alias{lor_bzr_sf}
\title{Geometry Bezirksregion}
\format{A data frame with 130 rows and 8 variables:
\describe{
\item{gml_id}{}
\item{RAUMID}{}
\item{BEZIRKSREGION}{}
\item{PROGNOSERAUM}{}
\item{BEZIRK}{}
\item{DATUM_GUELTIG_AB}{}
\item{FLAECHENGROESSE_IN_M2}{}
\item{geometry}{}
}}
\source{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_lor_bez@senstadt&type=WFS}
}
\usage{
lor_bzr_sf
}
\description{
An `sf` object containing the geometry and other attributes
}
\keyword{datasets}
