% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlingeo_data.R
\docType{data}
\name{lor_pgr_sf}
\alias{lor_pgr_sf}
\title{Geometry Prognoseraum}
\format{A data frame with 60 rows and 7 variables:
\describe{
  \item{gml_id}{}
  \item{RAUMID}{}
  \item{PGRNAME}{}
  \item{BEZNAME}{}
  \item{DATUM}{}
  \item{SHAPE_AREA}{}
  \item{geometry}{}
}}
\source{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_lor_prog@senstadt&type=WFS}
}
\usage{
lor_pgr_sf
}
\description{
An `sf` object containing the geometry and other attributes
}
\keyword{datasets}
