% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlingeo_data.R
\docType{data}
\name{lor_plr_sf}
\alias{lor_plr_sf}
\title{Geometry Planungsraum}
\format{A data frame with 447 rows and 9 variables:
\describe{
 \item{gml_id}{}
 \item{RAUMID}{}
 \item{PLANUNGSRAUM}{}
 \item{BEZIRKSREGION}{}
 \item{PROGNOSERAUM}{}
 \item{BEZIRK}{}
 \item{DATUM_GUELTIG_AB}{}
 \item{FLAECHENGROESSE_IN_M2}{}
 \item{geometry}{}
}}
\source{
\url{https://fbinter.stadt-berlin.de/fb/berlin/service_intern.jsp?id=s_lor_plan@senstadt&type=WFS}
}
\usage{
lor_plr_sf
}
\description{
An `sf` object containing the geometry and other attributes
}
\keyword{datasets}
